use json4s instead of hocon/configs to write json-state-files, move SnowparkSubFeed to sdl-snowflake
