+++ sdl-core/pom.xml
@@ -335,9 +335,9 @@
 		</dependency>
 
 		<dependency>
-			<groupId>com.snowflake</groupId>
-			<artifactId>snowpark</artifactId>
-			<version>${snowflake-snowpark.version}</version>
+			<groupId>org.json4s</groupId>
+			<artifactId>json4s-ext_${scala.minor.version}</artifactId>
+			<version>${json4s.version}</version>
 		</dependency>
 
 	</dependencies>
