+++ sdl-core/src/main/scala/io/smartdatalake/workflow/SubFeed.scala
@@ -18,10 +18,10 @@
  */
 package io.smartdatalake.workflow
 
-import com.snowflake.snowpark
+import io.smartdatalake.config.ParsableFromConfig
 import io.smartdatalake.config.SdlConfigObject.DataObjectId
 import io.smartdatalake.definitions.ExecutionModeResult
-import io.smartdatalake.util.dag.{DAG, DAGResult}
+import io.smartdatalake.util.dag.DAGResult
 import io.smartdatalake.util.hdfs.PartitionValues
 import io.smartdatalake.util.hive.HiveUtil
 import io.smartdatalake.util.misc.ScalaUtil.optionalizeMap
@@ -29,14 +29,16 @@ import io.smartdatalake.util.misc.{DataFrameUtil, SmartDataLakeLogger}
 import io.smartdatalake.util.streaming.DummyStreamProvider
 import io.smartdatalake.workflow.dataobject.FileRef
 import org.apache.spark.sql.types.StructType
-import org.apache.spark.sql.{Column, DataFrame, SparkSession, functions}
+import org.apache.spark.sql.{Column, DataFrame, functions}
 import sun.reflect.generics.reflectiveObjects.NotImplementedException
 
 /**
  * A SubFeed transports references to data between Actions.
  * Data can be represented by different technologies like Files or DataFrame.
+ *
+ * Note: SubFeed is implementing ParsableFromConfig to persist to
  */
-sealed trait SubFeed extends DAGResult with SmartDataLakeLogger {
+trait SubFeed extends DAGResult with SmartDataLakeLogger {
   def dataObjectId: DataObjectId
   def partitionValues: Seq[PartitionValues]
   def isDAGStart: Boolean
@@ -382,66 +384,5 @@ case class InitSubFeed(override val dataObjectId: DataObjectId,
   }
   def applyExecutionModeResultForOutput(result: ExecutionModeResult)(implicit context: ActionPipelineContext): SubFeed = {
     this.copy(partitionValues = result.inputPartitionValues, isSkipped = false)
-  }
-}
-
-case class SnowparkSubFeed(@transient dataFrame: Option[snowpark.DataFrame],
-                           override val dataObjectId: DataObjectId,
-                           override val partitionValues: Seq[PartitionValues],
-                           override val isDAGStart: Boolean = false,
-                           override val isSkipped: Boolean = false)
-  extends SubFeed {
-
-  override def clearPartitionValues(breakLineageOnChange: Boolean = true)(implicit context: ActionPipelineContext): SnowparkSubFeed = {
-    this.copy(partitionValues = Seq())
-  }
-
-  override def updatePartitionValues(partitions: Seq[String], breakLineageOnChange: Boolean = true, newPartitionValues: Option[Seq[PartitionValues]] = None)(implicit context: ActionPipelineContext): SnowparkSubFeed = {
-    val updatedPartitionValues = SubFeed.filterPartitionValues(newPartitionValues.getOrElse(partitionValues), partitions)
-    this.copy(partitionValues = updatedPartitionValues)
-  }
-
-  override def clearDAGStart(): SnowparkSubFeed = {
-    this.copy(isDAGStart = false)
-  }
-
-  override def clearSkipped(): SnowparkSubFeed = {
-    this.copy(isSkipped = false)
-  }
-
-  override def toOutput(dataObjectId: DataObjectId): SnowparkSubFeed = {
-    this.copy(dataFrame = None, isDAGStart = false, isSkipped = false, dataObjectId = dataObjectId)
-  }
-
-  override def union(other: SubFeed)(implicit context: ActionPipelineContext): SubFeed = other match {
-    case snowparkSubFeed: SnowparkSubFeed if this.dataFrame.isDefined && snowparkSubFeed.dataFrame.isDefined =>
-      this.copy(dataFrame = Some(this.dataFrame.get.unionByName(snowparkSubFeed.dataFrame.get)),
-        partitionValues = unionPartitionValues(snowparkSubFeed.partitionValues),
-        isDAGStart = this.isDAGStart || snowparkSubFeed.isDAGStart)
-    case subFeed =>
-      this.copy(dataFrame = None,
-        partitionValues = unionPartitionValues(subFeed.partitionValues),
-        isDAGStart = this.isDAGStart || subFeed.isDAGStart)
-  }
-
-  override def applyExecutionModeResultForInput(result: ExecutionModeResult, mainInputId: DataObjectId)(implicit context: ActionPipelineContext): SnowparkSubFeed = {
-    this.copy(partitionValues = result.inputPartitionValues, isSkipped = false)
-  }
-
-  override def breakLineage(implicit context: ActionPipelineContext): SnowparkSubFeed = {
-    this.copy()
-  }
-
-  override def applyExecutionModeResultForOutput(result: ExecutionModeResult)(implicit context: ActionPipelineContext): SnowparkSubFeed = {
-    this.copy(partitionValues = result.inputPartitionValues, isSkipped = false, dataFrame = None)
-  }
-}
-
-object SnowparkSubFeed extends SubFeedConverter[SnowparkSubFeed] {
-  override def fromSubFeed(subFeed: SubFeed)(implicit context: ActionPipelineContext): SnowparkSubFeed = {
-    subFeed match {
-      case snowparkSubFeed: SnowparkSubFeed => snowparkSubFeed
-      case _ => SnowparkSubFeed(None, subFeed.dataObjectId, subFeed.partitionValues, subFeed.isDAGStart, subFeed.isSkipped)
-    }
   }
 }
