+++ sdl-snowflake/pom.xml
@@ -82,6 +82,12 @@
 		</dependency>
 
 		<dependency>
+			<groupId>com.snowflake</groupId>
+			<artifactId>snowpark</artifactId>
+			<version>${snowflake-snowpark.version}</version>
+		</dependency>
+
+		<dependency>
 			<groupId>net.snowflake</groupId>
 			<artifactId>spark-snowflake_${scala.minor.version}</artifactId>
 			<version>${snowflake-spark.version}-spark_${spark.minor.version}</version>
